Port Collective Mind Improvements (#2297)
* port collective mind (#1847)

* port of a port

* port 699592f

* purge unrelated changes

* goobstation comments and move around new files

* add collective mind to things
misc features

* fix default radio

* actually fix

* namespac

* let ghosts hear all

---------


* collective mind improvement + port binary to collective mind (#2479)

* initial

* 500. prototype errors.

* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

* tide

* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

* i love webedits

* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

---------


* working

* chore: automatically update REUSE headers

* fix

* chore: automatically update REUSE headers

---------

+++ Content.Shared/Chat/ChatChannel.cs
@@ -11,6 +11,7 @@
 // SPDX-FileCopyrightText: 2023 Leon Friedrich
 // SPDX-FileCopyrightText: 2024 Julian Giebel
 // SPDX-FileCopyrightText: 2025 Ark
+// SPDX-FileCopyrightText: 2025 Ilya246
 // SPDX-FileCopyrightText: 2025 ark1368
 // SPDX-FileCopyrightText: 2025 starch
 //
@@ -22,7 +23,7 @@ namespace Content.Shared.Chat
     ///     Represents chat channels that the player can filter chat tabs by.
     /// </summary>
     [Flags]
-    public enum ChatChannel : ushort
+    public enum ChatChannel : uint // Goobstation - Starlight collective mind port // surely changing the ushort to uint won't break anything :clueless:
     {
         None = 0,
 
@@ -98,6 +99,7 @@ namespace Content.Shared.Chat
         /// </summary>
         AdminChat = 1 << 13,
 
+        // Goobstation - Starlight collective mind port
         /// <summary>
         ///     Collective mind channel for entities who have comp.
         /// </summary>
